+++ homeassistant/components/recorder/migration.py
@@ -54,15 +54,15 @@ from .queries import (
     find_states_context_ids_to_migrate,
     has_used_states_event_ids,
 )
-from .repairs.statistics import (
+from .repairs.statistics.duplicates import (
+    delete_statistics_duplicates,
+    delete_statistics_meta_duplicates,
+)
+from .repairs.statistics.schema import (
     correct_db_schema as statistics_correct_db_schema,
     validate_db_schema as statistics_validate_db_schema,
 )
-from .statistics import (
-    delete_statistics_duplicates,
-    delete_statistics_meta_duplicates,
-    get_start_time,
-)
+from .statistics import get_start_time
 from .tasks import (
     CommitTask,
     PostSchemaMigrationTask,
