+++ homeassistant/components/recorder/repairs/statistics/__init__.py
@@ -1,0 +1,1 @@
+"""Statistics repairs for Recorder."""
